Forgot to update references to renamed event
